Back out some CLI reorg
